Move EnforceSingleRowNode to spi
