+++ presto-main/src/main/java/com/facebook/presto/cost/EnforceSingleRowStatsRule.java
@@ -15,9 +15,9 @@ package com.facebook.presto.cost;
 
 import com.facebook.presto.Session;
 import com.facebook.presto.matching.Pattern;
+import com.facebook.presto.spi.plan.EnforceSingleRowNode;
 import com.facebook.presto.sql.planner.TypeProvider;
 import com.facebook.presto.sql.planner.iterative.Lookup;
-import com.facebook.presto.sql.planner.plan.EnforceSingleRowNode;
 
 import java.util.Optional;
