+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -79,4 +79,9 @@ public abstract class PlanVisitor<R, C>
     {
         return visitPlan(node, context);
     }
+
+    public R visitEnforceSingleRow(EnforceSingleRowNode node, C context)
+    {
+        return visitPlan(node, context);
+    }
 }
